+++ test/highs_tests.jl
@@ -239,7 +239,7 @@ if @isdefined HiGHS
         NemoMod.calculatescenario(dbfile; jumpmodel = (reg_jumpmode ? Model(HiGHS.Optimizer, add_bridges=false) : direct_model(HiGHS.Optimizer())),
             varstosave =
                 "vdemandnn, vnewcapacity, vtotalcapacityannual, vproductionbytechnologyannual, vproductionnn, vusebytechnologyannual, vusenn, vtotaldiscountedcost, "
-                * "vtransmissionbuilt, vtransmissionexists, vtransmissionbyline, vtransmissionannual", quiet = calculatescenario_quiet)
+                * "vtransmissionbuilt, vtransmissionexists, vtransmissionbyline, vtransmissionannual, vfuelprice, vfuelpricenodal, vfuelpriceannual, vfuelpricenodalannual", quiet = calculatescenario_quiet)
 
         db = SQLite.DB(dbfile)
 
@@ -267,6 +267,11 @@ if @isdefined HiGHS
             @test isapprox(testqry[8,:val], 178.714951753762; atol=TOL)
             @test isapprox(testqry[9,:val], 170.204748803116; atol=TOL)
             @test isapprox(testqry[10,:val], 162.099761418328; atol=TOL)
+
+            testqry = SQLite.DBInterface.execute(db, "select * from vfuelpriceannual where f = 'gas' order by y") |> DataFrame
+
+            @test testqry[1,:y] == "2020"
+            @test isapprox(testqry[1,:val], 150.0; atol=TOL)
         end
 
         # Test with non time sliced fuels